chore: update changes to fit style guide
