+++ lib/booleans/and/index.test.ts
@@ -52,8 +52,8 @@ test("returns an error when one or more operands is an error", async () => {
 	})()
 
 	expect(isLeft(failure)).toBeTruthy()
-	//expect((failure as Left<Array<string>>).left).toEqual([
-	//	"Unknown operation.",
-	//	"Unknown operation.",
-	//])
+	// expect((failure as Left<Array<string>>).left).toEqual([
+	// 	"Unknown operation.",
+	// 	"Unknown operation.",
+	// ])
 })
