+++ lib/booleans/and/index.ts
@@ -1,10 +1,11 @@
 import type { Either } from "fp-ts/lib/Either"
 import { left, right, traverseArray, match } from "fp-ts/lib/Either"
 import { pipe } from "fp-ts/lib/function"
+import { IO } from "fp-ts/lib/IO"
 
-import { composeOperations } from "../../operations/compose"
+import composeOperations from "../../operations/compose"
 
-type And = (op: AndOperation) => () => Either<Array<string>, boolean>
+type And = (op: AndOperation) => IO<Either<Array<string>, boolean>>
 const and: And = op => {
 	return pipe(
 		op.operands,
